requirements.txt
pandas
seaborn
matplotlib
python-pptx
loguru